Fixed tests for sync_repo_utils, upload_utils
