+++ lib/codesyncd/utils.js
@@ -6,7 +6,7 @@ import yaml from 'js-yaml';
 
 import { isBinaryFileSync } from "isbinaryfile";
 import { putLogEvent } from "../logger";
-import { uploadFileToServer } from "../utils/upload_file";
+import { uploadFileToServer } from "../utils/upload_utils";
 import { diff_match_patch } from "diff-match-patch";
 
 import { CONFIG_PATH, DIFF_SIZE_LIMIT, REQUIRED_DIFF_KEYS,
@@ -53,7 +53,8 @@ export const handleNewFileUpload = async (access_token, diffData, relPath, repoI
 			config: configJSON
 		};
 	}
-	const response = await uploadFileToServer(access_token, repoId, diffData.branch, originalsPath, relPath, diffData.created_at);
+	const response = await uploadFileToServer(access_token, repoId, diffData.branch, originalsPath,
+		relPath, diffData.created_at);
 	if (response.error) {
 		putLogEvent(`Error uploading new file to server: ${response.error}`);
 		return {
