+++ lib/init/utils.js
@@ -5,20 +5,18 @@ import path from 'path';
 import walk from 'walk';
 import yaml from 'js-yaml';
 import ignore from 'ignore';
-import fetch from "node-fetch";
 import parallel from "run-parallel";
 import getBranchName from "current-git-branch";
 import {isBinaryFileSync} from 'isbinaryfile';
 
 import {
-    API_INIT,
     NOTIFICATION,
     DEFAULT_BRANCH
 } from '../constants';
 import {getSkipRepos, getSyncIgnoreItems, isRepoActive, readYML} from '../utils/common';
 import {checkServerDown} from '../utils/api_utils';
 import {putLogEvent} from '../logger';
-import {uploadFileTos3} from '../utils/upload_file';
+import {uploadFileTos3, uploadRepoToServer} from '../utils/upload_utils';
 import {trackRepoHandler, unSyncHandler} from '../handlers/commands_handler';
 import {generateSettings} from "../settings";
 
@@ -120,27 +118,6 @@ export class initUtils {
             }
         });
     }
-
-    static uploadRepoToServer = async (token, data) => {
-        let error = '';
-        const response = await fetch(API_INIT, {
-                method: 'post',
-                body: JSON.stringify(data),
-                headers: {
-                    'Content-Type': 'application/json',
-                    'Authorization': `Basic ${token}`
-                },
-            }
-        )
-            .then(res => res.json())
-            .then(json => json)
-            .catch(err => error = err);
-
-        return {
-            response,
-            error
-        };
-    };
 
     static saveIamUser(user) {
         // save iam credentials if not saved already
@@ -280,7 +257,7 @@ export class initUtils {
             branch,
             files_data: JSON.stringify(filesData)
         };
-        const json = await initUtils.uploadRepoToServer(token, data);
+        const json = await uploadRepoToServer(token, data);
         if (json.error || json.response.error) {
 			const error = isSyncingBranch ? NOTIFICATION.ERROR_SYNCING_BRANCH : NOTIFICATION.ERROR_SYNCING_REPO;
 			putLogEvent(`${error}. Reason: ${json.error || json.response.error}`);
