+++ lib/utils/sync_repo_utils.js
@@ -6,7 +6,7 @@ import { API_ENDPOINT } from "../constants";
 
 export const updateRepo = async (accessToken, repoId, data) => {
     let error = "";
-    const response = await fetch(`${API_ENDPOINT}/repos/${repoId}`, {
+    let response = await fetch(`${API_ENDPOINT}/repos/${repoId}`, {
         method: 'patch',
         body: JSON.stringify(data),
         headers: {
@@ -17,6 +17,13 @@ export const updateRepo = async (accessToken, repoId, data) => {
         .then(res => res.json())
         .then(json => json)
         .catch(err => error = err);
+
+    if ("error" in response) {
+        error = response["error"];
+    }
+    if (error) {
+        response = {};
+    }
 
     return {
         response,
